image_processor/requirements.txt
boto3
pillow 